+++ gui.hpp
@@ -149,7 +149,7 @@ public:
         if (n2 != NULL && n1 != NULL && n1 != n2){
             //have gui ask for node link weight
             std::cout << "4- Linking nodes\n";
-            int link_weight = -1;
+            ll link_weight = -1;
 
             //std::string title = "Set link weight for nodes" + std::to_string(n1->getNodeIdent()) + " and " + std::to_string(n2->getNodeIdent());
             graphMan->joinNodes(n1, n2, 20, lstate);
